gdb: arch to tell whether it supports 'g' packets
This adds a callback to gdbarch so indicate whether
the architecture supports 'g' packets to remote target.
For example, on Xtensa, xt-ocd does not support 'g'
packet. However, the register fetching code keeps on
trying 'g' packet which results in infinite loop. So
this provides a mechanism for architecture to indicate
whether 'g' packet is supported.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>