objcopy: Don't enforce alignment when adjusting LMA
Allow the first section LMA to match either the raw LMA of the
segment or the alignment-adjusted LMA of the segment. I'm not sure why
the LMA should ever be adjusted with the VMA alignment, but this change
doesn't affect the output when things work, it only makes objcopy
less strict.

Signed-off-by: Keith Packard <keithp@keithp.com>